requirements.txt
pyyaml==6.0.1
uuid==1.30
psutil==5.9.8